+++ src/components/cc-tcp-redirection-form/cc-tcp-redirection-form.types.d.ts
@@ -1,10 +1,1 @@
-export interface RedirectionNamespace {
-  namespace: string,
-}
-
-export interface Redirection {
-  namespace: string,
-  sourcePort: number,
-}
-
-export type ContextRedirectionType = "user" | "admin";
+type ContextRedirectionType = "user" | "admin";