+++ src/components/cc-toaster/cc-toaster.js
@@ -6,9 +6,7 @@ import '../cc-toast/cc-toast.js';
 /**
  * @typedef {import('./cc-toaster.types.js').ToastPosition} ToastPosition
  * @typedef {import('./cc-toaster.types.js').ToastAnimation} ToastAnimation
- * @typedef {import('./cc-toaster.types.js').Notification} Notification
  * @typedef {import('./cc-toaster.types.js').ToastOptions} ToastOptions
- * @typedef {import('./cc-toaster.types.js').Toast} Toast
  */
 
 /**